Use proper type so we don't have to cast it.

+++ src/savegame_json.cpp
@@ -1826,9 +1826,9 @@ void monster::load( const JsonObject &data )
     // sp_timeout indicates an old save, prior to the special_attacks refactor
     if( data.has_array( "sp_timeout" ) ) {
         JsonArray parray = data.get_array( "sp_timeout" );
-        int index = 0;
+        size_t index = 0;
         int ptimeout = 0;
-        while( parray.has_more() && index < static_cast<int>( type->special_attacks_names.size() ) ) {
+        while( parray.has_more() && index < type->special_attacks_names.size() ) {
             if( parray.read_next( ptimeout ) ) {
                 // assume timeouts saved in same order as current monsters.json listing
                 const std::string &aname = type->special_attacks_names[index++];
